+++ Mods ChoGGi/Add Progress To Training Section/metadata.lua
@@ -1,0 +1,18 @@
+return PlaceObj("ModDef", {
+	"title", "Add Progress To Training Section",
+	"version_major", 0,
+	"version_minor", 1,
+	"saved", 1555329600,
+	"image", "Preview.png",
+	"id", "ChoGGi_AddProgressToTrainingSection",
+	"steam_id", "1713299705",
+	"pops_any_uuid", "00b83936-ba0f-4823-8e65-a16e981e6861",
+	"author", "ChoGGi",
+	"lua_revision", LuaRevision or 244124,
+	"code", {
+		"Code/Script.lua",
+	},
+	"description", [[Edits tooltip to show list of students and their training progress.
+
+Requested by Tassyr.]],
+})