+++ Mods ChoGGi/Fix Disable Fireworks/Code/Script.lua
@@ -1,0 +1,6 @@
+local function StartupCode()
+	hr.ShowFireworks = 0
+end
+
+OnMsg.CityStart = StartupCode
+OnMsg.LoadGame = StartupCode